# backend/bait-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API in Laravel",
        "description": "API documentation with Swagger in Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/profile/show": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get the authenticated user's profile",
                "description": "Returns the authenticated user's profile with related role, state, avatar, and banner information.",
                "operationId": "05c552df192df4e9f92936e798a9e15f",
                "responses": {
                    "200": {
                        "description": "User profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/update": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update the authenticated user's profile",
                "description": "Updates the authenticated user's profile. Only 'name', 'username', and 'email' can be changed.",
                "operationId": "514d6028989771b7cbaa85cda3d66cc2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "SuperAdmin"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Super Admin"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "superadmin@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/password": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Change the authenticated user's password",
                "description": "Allows the authenticated user to change their password by providing the current and new passwords.",
                "operationId": "3b72cc50e5174562a87320c942b7d065",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "description": "The user's current password",
                                        "type": "string",
                                        "format": "password",
                                        "example": "admin123"
                                    },
                                    "new_password": {
                                        "description": "The new password the user wants to set",
                                        "type": "string",
                                        "format": "password",
                                        "example": "superadmin123"
                                    },
                                    "new_password_confirmation": {
                                        "description": "Confirmation of the new password",
                                        "type": "string",
                                        "format": "password",
                                        "example": "superadmin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Password updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated"
                    },
                    "422": {
                        "description": "Validation failed (e.g., current password incorrect, password too weak)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/privileged/users/{user}/update": {
            "put": {
                "tags": [
                    "User Management"
                ],
                "summary": "Update a user's profile",
                "description": "Allows admins and moderators to update a user's profile. Moderators can only update limited fields (e.g., name).",
                "operationId": "ea815b53cffaaca971c03382641fdeba",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "janedoe"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane@example.com"
                                    },
                                    "state_id": {
                                        "description": "ID of the new state for the user",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "role_id": {
                                        "description": "ID of the new role for the user",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSchema"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - insufficient permissions"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/privileged/users/{user}/password": {
            "put": {
                "tags": [
                    "User Management"
                ],
                "summary": "Change password for a specified user *ADMIN Only",
                "description": "Allows admins or moderators to update the password of a specified user.",
                "operationId": "727402a3c29eb8a371f35333ec748863",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user whose password is to be changed",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "new_password"
                                ],
                                "properties": {
                                    "new_password": {
                                        "description": "The new password for the user (min 8 characters)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "newStrongPassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Password for user johndoe updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - insufficient permissions"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation error - e.g., password too short"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AvatarSchema": {
                "title": "Avatar Schema",
                "required": [
                    "id",
                    "url_avatars"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "url_avatars": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/storage/avatars/avatar1.jpg"
                    }
                },
                "type": "object"
            },
            "BannerSchema": {
                "title": "Banner Schema",
                "required": [
                    "id",
                    "url_banners"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "url_banners": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/storage/banners/banner1.jpg"
                    }
                },
                "type": "object"
            },
...
            "UserSchema": {
                "title": "User Schema",
                "required": [
                    "id",
                    "username",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 123
                    },
                    "username": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "role": {
                        "description": "User's role name",
                        "type": "string",
                        "example": "admin",
                        "nullable": true
                    },
                    "state": {
                        "description": "User's state name",
                        "type": "string",
                        "example": "active",
                        "nullable": true
                    },
                    "avatar": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/AvatarSchema"
                            }
                        ],
                        "nullable": true
                    },
                    "banner": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/BannerSchema"
                            }
                        ],
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-05T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-06T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a JWT token to authenticate",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}